backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
python-multipart==0.0.20
requests==2.32.3
pytest==9.0.2
httpx==0.28.1
Pillow>=10.0.0
numpy>=1.24.0
